refact: mudando metódo que busca perguntas de lugar, para fazer mais sentido
