+++ src/models/Projeto.php
@@ -522,34 +522,6 @@ class Projeto
         }
     }
 
-    public static function buscaPerguntas()
-    {
-        global $conn;
-        try {
-            $sql = "SELECT id_pergunta, texto_pergunta, tipo_pergunta, ordem FROM pergunta";
-            $stmt = $conn->prepare($sql);
 
-            if (!$stmt) {
-                throw new Exception("Erro ao preparar a query: " . $conn->error);
-            }
-
-            if (!$stmt->execute()) {
-                throw new Exception("Erro ao executar a query: " . $stmt->error);
-            }
-
-            $resultado = $stmt->get_result();
-            $perguntas = [];
-
-            while ($row = $resultado->fetch_assoc()) {
-                $perguntas[] = $row;
-            }
-
-            return $perguntas;
-
-        } catch (Exception $e) {
-            error_log($e->getMessage()); // Registra o erro no log do servidor
-            return ['erro' => 'Não foi possível buscar as perguntas.']; // Retorna um erro amigável
-        }
-    }
 
 }
